move SaveItems into module tools
it was defined exactly the same way in two files in the same directory
