+++ smoderp2d/tools/save_load_data_nopickle.py
@@ -10,51 +10,7 @@ import os
 import sys
 import zipfile
 
-
-# Class to save item of different typy
-#
-class SaveItems:
-
-    def savelist(self, l):
-        a = 0
-        b = []
-        self.f.writelines(str(len(l)) + '\n')
-        for i in range(len(l)):
-            if not l[i]:
-                pass
-            else:
-                if isinstance(l[i], list):
-                    for j in range(len(l[i])):
-                        b.append([a, l[i][j]])
-                else:
-                    b.append([a, l[i]])
-            a += 1
-        for item1 in b:
-            line = ''
-            for item2 in item1:
-                line += str(item2) + ';'
-            line = line[:-1]
-            self.f.writelines(line + '\n')
-
-    def saveint(self, f):
-        self.f.writelines(str(f) + '\n')
-
-    def savefloat(self, f):
-        self.f.writelines(str(f) + '\n')
-
-    def savestr(self, s):
-        self.f.writelines(s + '\n')
-
-    def saveunicode(self, uni):
-        self.f.writelines(uni + '\n')
-
-    def savenumpy(self, npa):
-        type_ = str(type(npa[0][0]))
-        self.f.writelines(type_ + '\n')
-        if 'int' in type_:
-            np.savetxt(self.f, npa, fmt='%15d', delimiter=';')
-        if 'float' in type_:
-            np.savetxt(self.f, npa, fmt='%15.10e', delimiter=';')
+from tools import SaveItems
 
 
 # Class to load item of different typy
